@@ -4,10 +4,11 @@ classifiers =
     Intended Audience :: Science/Research
     Topic :: Scientific/Engineering :: Image Recognition
     License :: OSI Approved :: Apache Software License
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
+    Programming Language :: Python :: 3.12
 description = TemplateFlow Python Client - TemplateFlow is the Zone of neuroimaging templates.
 license = Apache-2.0
 license_file = LICENSE
@@ -24,13 +25,14 @@ project_urls =
     Source Code = https://github.com/templateflow/python-client
 
 [options]
-python_requires = >= 3.7
+python_requires = >= 3.8
 setup_requires =
     setuptools >= 45
     setuptools_scm >= 6.2
     wheel
 install_requires =
     pybids >= 0.15.2
+    importlib_resources >= 5.7; python_version < '3.11'
     requests
     tqdm
 test_requires =
